competition/requirements.txt
python-chess
chess
pyaml
joblib
more_itertools
matplotlib
pandas